Fix package name starting with number
