+++ scripts/src/lib/template/java.ts
@@ -96,3 +96,44 @@ export function computePackageNameErrors(packageName: string): string[] {
 
 	return errorList;
 }
+
+export function formatPackageName(packageName: string): string {
+    return packageName
+        .toLocaleLowerCase()
+        .replaceAll(/\s+/g, '.')
+        .replaceAll(/[^a-za-z0-9_\.]/g, "");
+}
+
+export function generatePackageName(packageName: string): string | undefined {
+    const formatted = replaceNumerals(formatPackageName(packageName)).join('');
+
+    if(computePackageNameErrors(formatted).length !== 0)
+        return;
+
+    return formatted;
+}
+
+const NUMERALS = [
+    'zero',
+    'one',
+    'two',
+    'three',
+    'four',
+    'five',
+    'six',
+    'seven',
+    'eight',
+    'nine'
+] as const;
+
+export function replaceNumerals(s: string): string[] {
+    const match = s.match(/^\d+/);
+    if(!(match && match[0]))
+        return [s];
+    const numericPrefix = match[0];
+    const rest = s.substring(numericPrefix.length);
+    const out: string[] = numericPrefix.split('').map(x => NUMERALS[parseInt(x)]);
+    if(rest)
+        out.push(rest);
+    return out;
+}
